+++ radical_translations/core/serializers.py
@@ -15,3 +15,9 @@ class ResourceDocumentSerializer(DocumentSerializer):
             return obj.meta.highlight.__dict__["_d_"]
 
         return {}
+
+
+class SimpleResourceDocumentSerializer(DocumentSerializer):
+    class Meta:
+        document = ResourceDocument
+        fields = ["id", "year", "places"]