+++ CHANGELOG.rst
@@ -32,6 +32,7 @@ Added
 * Docs: data model fields concordance.
 * Command to export `Places` to a CSV file.
 * Convert notes and summary URLs into clickable links.
+* Badge for related resources that have a paratext.
 
 Changed
 ~~~~~~~
@@ -47,6 +48,7 @@ Changed
 * Order/display of `Contributions`.
 * Labels and descriptions of the classification vocabularies.
 
+
 Fixed
 ~~~~~
 * Validation errors in `CHANGELOG`.
@@ -59,6 +61,8 @@ Fixed
 Removed
 ~~~~~~~
 * `Place` detail view.
+* Radical marker from dates.
+* Radical markers filter from the search.
 
 .. _profiles: https://docs.docker.com/compose/profiles/
 .. _Bump pillow to 7.1: https://github.com/kingsdigitallab/radical_translations/pull/5